+++ prisma/prod/migrations/4_total_starters/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Regatta" ADD COLUMN "totalStarters" INTEGER;